stock-prediction-project/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
scrapy==2.11.0
selenium==4.15.2
lxml==4.9.3
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.3.2
python-dotenv==1.0.0
tensorflow==2.15.0
keras==2.15.0
xgboost==2.0.2
lightgbm==4.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pymysql==1.1.0
fastapi==0.104.1
uvicorn==0.24.0
aiofiles==23.2.1
celery==5.3.4
apscheduler==3.10.4
pytest==7.4.3
pytest-asyncio==0.21.1